pyproject.toml
[tool.poetry]
name = "cpmbot"
version = "0.1.0"
description = ""
authors = ["Alexander Averyanov <alex@averyan.ru>"]

[tool.poetry.scripts]
cpmbot = 'cpmbot:main'

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.0"
requests = "^2.28.2"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
